requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
alembic==1.13.1
aiosqlite==0.19.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Google Drive
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Azure Blob Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.2

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Logging
loguru==0.7.2
